+++ .gitignore
@@ -129,3 +129,6 @@ dmypy.json
 # Pyre type checker
 .pyre/
 .tmp.driveupload*
+
+#Random
+.idea/
